delegate push_lock_wait
